修改映射文件，以及首页业务逻辑

+++ src/main/java/com/tao/mapper/UserInfoDao.java
@@ -14,7 +14,7 @@ public interface UserInfoDao {
 
     List<UserInfo> findAllUserInfo();
 
-    UserInfo findUserInfoById(int id);
+    UserInfo findUserInfoById(long id);
 
     void insertUserInfo(UserInfo userInfo);
 }
